// src/app/models/Eventos.ts
export interface RegistroActividadesEvento{
    Id_Actividad?: number;
    Nombre_Actividad?: string;
    Fecha_Actividad?: Date;
    Hora_Inicio?: string;
    Hora_Termino?: string;
    Cantidad_Asistentes?: number;
    Lugar?: string;
    id_TipoActividad?: number;
    Nombre_Responsable?: string;
    Nombre_Participante?: string;
    
}